+++ crates/http-source/src/config.rs
@@ -58,16 +58,8 @@ pub(crate) struct HttpConfig {
 #[derive(Debug)]
 pub(crate) struct WebSocketConfig {
     pub(crate) subscription_message: Option<String>,
-    pub(crate) reconnection_policy: Option<ReconnectionPolicy>,
     // TODO: pub(crate) max_message_size: Option<usize>,
     pub(crate) ping_interval_ms: Option<u64>,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub(crate) struct ReconnectionPolicy {
-    pub(crate) max_retries: u32,
-    pub(crate) base_delay_ms: usize,
-    pub(crate) max_delay_ms: usize,
 }
 
 #[derive(Debug, Default, Deserialize, Clone, Copy)]
